Modularizing TKG test
- fixes #445
